[INTERPRETER] * BUG: Fix the custom realloc() internal routine on 64 bits systems.
[COMPILER]
* BUG: Fix the custom realloc() internal routine on 64 bits systems.


git-svn-id: svn://localhost/gambas/trunk@3021 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_alloc_temp.h
@@ -36,6 +36,8 @@
 
 int MEMORY_count = 0;
 
+//#define DEBUG_ME
+
 #if DEBUG_MEMORY
 
 #include "gb_error.h"
@@ -359,7 +361,7 @@ void *my_realloc(void *alloc, size_t new_len)
 	ptr = alloc;
 	ptr--;
 	size = (int)*ptr;
-	new_size = REAL_SIZE(new_len + sizeof(int));
+	new_size = REAL_SIZE(new_len + sizeof(size_t));
 	
 	if (size == new_size)
 		return alloc;
